Motor speed control now works; Beginning of PID error correction

+++ src/Encoder.cpp
@@ -33,7 +33,7 @@ Encoder::~Encoder() {}
  * @return tick number
  */
 
-long Encoder::getTicks()
+double Encoder::getTicks()
 {
     return this->tickNumber;
 }
